Update main_project_queries.sql
Goal:

Get details of all the teachers who have been infected by COVID-19 in the
past two weeks. Details include the teacher’s first-name, last-name, date of
infection, and the name of the facility that the teacher is currently teaching for.
Results should be displayed sorted in ascending order by the facility name,
then by the first name of the teacher.

Previous change was not filtering covid 19 and only 2 weeks of difference.